refactor(indexer): standardize report field names and improve error handling in environment variable validation
